requirements.txt
# =============================================================================
# Superset MCP - Requirements
# =============================================================================

# -----------------------------------------------------------------------------
# Core MCP Framework
# -----------------------------------------------------------------------------
fastmcp>=0.2.0              # MCP server/client framework

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx>=0.24.0               # Async HTTP client for Superset API calls

# -----------------------------------------------------------------------------
# Web Framework (Chatbot API)
# -----------------------------------------------------------------------------
flask>=2.3.0                # Web framework for chatbot API server
flask-cors>=4.0.0           # CORS support for cross-origin requests

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
pyyaml>=6.0                 # YAML parsing for relationships config
python-dotenv>=1.0.0        # Environment variable loading from .env files

# -----------------------------------------------------------------------------
# LLM Clients
# -----------------------------------------------------------------------------
openai>=1.0.0               # OpenAI API for SQL generation
anthropic>=0.18.0           # Anthropic Claude API for SQL generation
google-genai>=0.3.0         # Google Gemini API for SQL critique/improvement

# -----------------------------------------------------------------------------
# NLU / Fuzzy Matching
# -----------------------------------------------------------------------------
rapidfuzz>=3.0.0            # Fast fuzzy string matching for semantic matching

# -----------------------------------------------------------------------------
# SQL Processing
# -----------------------------------------------------------------------------
sqlparse>=0.4.4             # SQL parsing for validation and analysis
sqlglot>=20.0.0             # Dialect-aware SQL parsing (StarRocks, PostgreSQL)

# -----------------------------------------------------------------------------
# Database Adapters
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9.0      # PostgreSQL adapter (for knowledge store)


# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.0.0               # Testing framework

# -----------------------------------------------------------------------------
# Optional Dependencies
# -----------------------------------------------------------------------------
# Uncomment as needed:
# redis>=4.0.0              # For caching (if using Redis cache)
# celery>=5.3.0             # For async task processing